@@ -1,6 +1,6 @@
 {
-  "name": "react-edit-inline",
-  "version": "1.0.8",
+  "name": "react-edit-inplace",
+  "version": "1.0.9",
   "description": "A simple inline text editor for React with ECMAScript 6 + JSX Harmony syntax",
   "main": "index.js",
   "scripts": {
@@ -11,7 +11,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/kaivi/ReactInlineEdit.git"
+    "url": "git+https://github.com/dkwingsmt/ReactInlineEdit.git"
   },
   "keywords": [
     "react",
@@ -21,14 +21,14 @@
     "input",
     "form"
   ],
-  "author": "Kai Vik <kvi@ieee.org>",
+  "author": "DKWings <dkwingsmt@gmail.com>",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/kaivi/ReactInlineEdit/issues"
+    "url": "https://github.com/dkwingsmt/ReactInlineEdit/issues"
   },
-  "homepage": "https://github.com/kaivi/ReactInlineEdit#readme",
+  "homepage": "https://github.com/dkwingsmt/ReactInlineEdit#readme",
   "peerDependencies": {
-    "react": "15.5.x"
+    "react": "16.x || 15.5.x"
   },
   "devDependencies": {
     "babel-cli": "^6.4.0",
@@ -40,6 +40,7 @@
     "webpack": "^1.12.11"
   },
   "dependencies": {
-    "prop-types": "^15.5.x"
+    "prop-types": "^15.6",
+    "react": "^16.0.0"
   }
 }
